pyproject.toml
[project]
name = "CIVISAIPolicyAnalyserBackend"
version = "1.0.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.11.3",
    "fastapi==0.115.11",
    "fastapi-cli==0.0.7",
    "FastAPI-SQLAlchemy==0.2.1",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "uvicorn>=0.34.2",
    "loguru>=0.7.3",
    "langchain>=0.3.25",
    "langchain-deepseek>=0.1.3",
    "pymupdf>=1.26.1",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "coverage==7.8.0",
    "pytest-cov==6.1.1",
    "ruff>=0.11.6",
]


[tool.ruff.lint]
ignore-init-module-imports = false
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
ignore = [
    "E501"  # Line is Too Long
]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
# Assume Python >3.9
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# 5. Use single quotes in `ruff format`.
quote-style = "single"

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.isort]
case-sensitive = true
combine-as-imports = true
detect-same-package = false

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F403"]
"src/civis_backend_policy_analyser/alembic/*" = ["UP035", "I001", "UP007"]
